Initial working version of gitdeploy
gitdeploy is a small HTTP server written in Python. Listening
to a given port, gitdeploy can receive requests from popular
git repositories such as GitHub, GitLab or Bitbucket and
automatically update local git repositories. It can further
follow a set of user-defined rules to deploy the given
repositories.

gitdeploy aims to be usable by the different users of the
system it is running on, and not only by the administrators or
that system. Each user can thus define its set of rules that
will be analyzed by gitdeploy and ran as this user, preventing
any unauthorized commands to run.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>