Add 'event' parameter for the user
The 'event' parameter answers the same way as the 'tag' and
'branch' ones, using 'only' and 'except' to specify which kind
of events to answer to. The basic event one is 'push'.

The pull workflow will also be done only if it is appropriate,
i.e. only if the event is a 'push' or 'create'.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>