Add a security key to run the deploy commands
The parameter 'key' can be added to each rule to specify
a key to match before running the deploy commands. The
webserver has to receive a request with the query string
key=value that will match value against the key in the rule.

For instance, for a local server on port 4242, the web hook
url would be: http://localhost:4242/?key=value

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>