Add ability to use secret hash in request headers
GitHub allows users to receive a header 'X-Hub-Signature' that
contains the content of the request hashed in sha1 with a salt
defined by the user. It is now possible to match this hash by
setting a 'secret' parameter in the rule with value the secret
key used to salt the hash.

This security check also works for different kind of hashes
that are available in hashlib, such as md5 or sha512. The
relevant method will be choosen when reading the
X-...-Signature header. This has been made in hope that some
other platform will support that kind of security, way better
than the query string check.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>